Support for pause/resume.
BUG=chromium-os:23516
TEST=Pause a download, upload should pause. Same for resume.
Review URL: https://chromiumcodereview.appspot.com/9666047

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@126315 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c2e3fc490e25de1270f806f27ff89cb767a75c60

+++ chrome/browser/chromeos/gdata/gdata_uploader.h
@@ -31,7 +31,7 @@ class GDataUploader {
 
   // Updates size and download_completed of streaming upload.
   void UpdateUpload(const GURL& file_url,
-                    size_t file_size,
+                    int64 file_size,
                     bool download_complete);
 
  private:
